api/requirements.txt
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
requests==2.32.3
python-keycloak==4.5.0
PyJWT==2.9.0
cryptography==44.0.0
httpx==0.27.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
Flask-SQLAlchemy==3.1.1
